requirements.txt
opencv-python>=4.12.0.88
ultralytics>=8.3.235
python-dotenv>=1.0.0
langchain>=0.3.27
langchain-openai>=0.3.35
langchain-anthropic>=0.1.0
langsmith>=0.4.33
rapidfuzz>=3.14.3
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
Pillow>=9.5.0
